Move linebreakless tag into aidants_connect common package
