+++ aidants_connect/common/tests/test_templatetags.py
@@ -1,5 +1,7 @@
-from unittest import TestCase
 from urllib.parse import quote
+
+from django.template import Context, Template
+from django.test import TestCase
 
 from aidants_connect.templatetags.ac_common import mailto
 
@@ -15,3 +17,68 @@ class Test(TestCase):
             f'<a href="mailto:{recipient}?subject={quote(subject, "")}&'
             f'body={quote(body, "")}">{mail_link}</a>',
         )
+
+
+class LinebreaklessTests(TestCase):
+    def test_linebreaks_are_stripped(self):
+        result = self._render_template(
+            """{% load ac_common %}
+            {% linebreakless %}
+            <p>
+                This is a test with a list
+            </p>
+            <ul>
+                <li>item 1</li>
+                <li>item 2</li>
+                <li>item 3</li>
+            </ul>
+            {% endlinebreakless %}"""
+        )
+
+        self.assertEqual(
+            "<p>This is a test with a list</p><ul><li>item 1</li><li>item 2</li>"
+            "<li>item 3</li></ul>",
+            result.strip(),
+        )
+
+    def test_keeplinebreak_keeps_linebreaks(self):
+        result = self._render_template(
+            """
+            {% load ac_common %}
+            {% linebreakless %}
+            <p>
+                This is a test with a list
+            </p>
+            <ul>{% keeplinebreak %}
+                <li>item 1</li>{% keeplinebreak %}
+                <li>item 2</li>{% keeplinebreak %}
+                <li>item 3</li>{% keeplinebreak %}
+            </ul>
+            {% endlinebreakless %}"""
+        )
+
+        self.assertEqual(
+            "<p>This is a test with a list</p><ul>\n<li>item 1</li>\n<li>item 2</li>\n"
+            "<li>item 3</li>\n</ul>",
+            result.strip(),
+        )
+
+    def test_keeplinebreak_throws_expception_when_not_followed_by_a_linebreak(self):
+        template = """
+            {% load ac_common %}
+            {% linebreakless %}
+            <p>
+                This is a test with a list
+            </p>
+            <ul>
+                <li>item 1</li>
+                <li>item 2</li>
+                <li>item 3</li>
+            </ul>
+            {% keeplinebreak %}{% endlinebreakless %}"""
+
+        with self.assertRaises(ValueError):
+            self._render_template(template)
+
+    def _render_template(self, string, context=None):
+        return Template(string).render(Context(context or {}))
