+++ aidants_connect_web/templates/aidants_connect_web/espace_responsable/confirm-remove-aidant-from-organisation.html
@@ -1,6 +1,6 @@
 {% extends 'layouts/main.html' %}
 
-{% load static ac_extras %}
+{% load static ac_extras ac_common %}
 
 {% linebreakless %}
   {% block title %}
