+++ aidants_connect_web/templates/aidants_connect_web/managment/notify_no_totp_workers.html
@@ -1,5 +1,5 @@
 {% extends "layouts/email_base.html" %}
-{% load ac_extras %}
+{% load ac_extras ac_common %}
 
 {% block email_body %}
   <table border="0" cellpadding="0" cellspacing="0" class="body">
