+++ aidants_connect_web/templates/aidants_connect_web/managment/notify_no_totp_workers.txt
@@ -1,4 +1,4 @@
-{% load ac_extras %}
+{% load ac_extras ac_common %}
 Bonjour,
 {% if not users  %}
 L’équipe Aidants Connect vous informe qu'aucune carte Aidants Connect n'est liée à votre compte.
