+++ aidants_connect_web/templates/signals/aidant__organisations_changed.txt
@@ -1,4 +1,4 @@
-{% load ac_extras %}
+{% load ac_extras ac_common %}
 
 Bonjour,
