+++ aidants_connect_web/templatetags/ac_extras.py
@@ -1,9 +1,6 @@
 import json
-from re import sub as re_sub
 
 from django import template
-from django.template import Library
-from django.template.base import Node, NodeList, Parser, TextNode, Token
 
 register = template.Library()
 
@@ -63,52 +60,3 @@ def list_term(context, **kwargs):
     additionnal_cond = kwargs.get("additionnal_cond", False)
 
     return end_term if forloop["last"] and not additionnal_cond else more_term
-
-
-@register.tag
-def linebreakless(parser: Parser, token: Token):
-    library = Library()
-    library.tag(LinebreaklessNode.KEEP_LINEBREAK_TAG, LinebreaklessNode.keeplinebreak)
-    parser.add_library(library)
-    nodelist = parser.parse(("endlinebreakless",))
-    parser.delete_first_token()
-    return LinebreaklessNode(token, nodelist)
-
-
-class LinebreaklessNode(Node):
-    KEEP_LINEBREAK_TAG = "keeplinebreak"
-    KEEP_LINEBREAK_MARKUP = f"{{% {KEEP_LINEBREAK_TAG} %}}"
-
-    class KeepLineBreak(Node):
-        def __init__(self, token: Token):
-            self.token = token
-
-        def render(self, context):
-            return LinebreaklessNode.KEEP_LINEBREAK_MARKUP
-
-    @staticmethod
-    def keeplinebreak(_, token: Token):
-        return LinebreaklessNode.KeepLineBreak(token)
-
-    def __init__(self, token: Token, nodelist: NodeList):
-        self.token = token
-        self.nodelist = nodelist
-
-    def render(self, context):
-        for i, node in enumerate(self.nodelist):
-            if isinstance(node, LinebreaklessNode.KeepLineBreak) and (
-                len(self.nodelist) <= i + 1
-                or not isinstance(self.nodelist[i + 1], TextNode)
-                or not self.nodelist[i + 1].s.strip(" \t\r\f\v").startswith("\n")
-            ):
-                raise ValueError(
-                    f"{LinebreaklessNode.KEEP_LINEBREAK_MARKUP} needs to be followed "
-                    "by a linebreak.\n"
-                    f'  File "{node.origin.name}", line {node.token.lineno}'
-                )
-
-        return re_sub(
-            "\\s*\n+\\s*",
-            "",
-            self.nodelist.render(context).strip(),
-        ).replace(self.KEEP_LINEBREAK_MARKUP, "\n")
